pyproject.toml
[project]
name = "gemini-mcp-tools"
version = "0.1.0"
description = "Repo for Research about Gemini x MCP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.10.0",
    "mcp>=1.6.0",
]
